# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Solve Corporate CRM API
  version: 1.0.0
  description: API do CRM Solve Corporate - Ecossistema Digital Integrado
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [novo_lead, contacto, qualificado, proposta, negociacao, convertido, perdido]
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  total:
                    type: integer
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 2
                email:
                  type: string
                  format: email
                phone:
                  type: string
                company:
                  type: string
                source:
                  type: string
                status:
                  type: string
                  enum: [novo_lead, contacto, qualificado, proposta, negociacao, convertido, perdido]
                  default: novo_lead
                ownerId:
                  type: string
                  format: uuid
                estimatedValue:
                  type: integer
                notes:
                  type: string
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Lead deleted
        "404":
          description: Lead not found
  /leads/{id}/convert:
    post:
      operationId: convertLead
      tags: [leads]
      summary: Convert lead to customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead converted
        "400":
          description: Lead already converted
        "404":
          description: Lead not found
...
  /integrations:
    get:
      operationId: listIntegrations
      tags: [integrations]
      summary: List all integrations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Integrations list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Integration"
                  total:
                    type: integer
  /integrations/{name}:
    get:
      operationId: getIntegration
      tags: [integrations]
      summary: Get integration by name
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Integration details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Integration"
        "404":
          description: Integration not found
    patch:
      operationId: updateIntegration
      tags: [integrations]
      summary: Update integration config
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [operacional, atencao, erro, inativo]
                config:
                  type: object
      responses:
        "200":
          description: Integration updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Integration"
        "404":
          description: Integration not found
  /integrations/{name}/sync:
    post:
      operationId: syncIntegration
      tags: [integrations]
      summary: Sync integration
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Integration"
                  message:
                    type: string
        "404":
          description: Integration not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [novo_lead, contacto, qualificado, proposta, negociacao, convertido, perdido]
        ownerId:
          type: string
          format: uuid
        estimatedValue:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [novo_lead, contacto, qualificado, proposta, negociacao, convertido, perdido]
        ownerId:
          type: string
          format: uuid
        estimatedValue:
          type: integer
        notes:
          type: string
...
    Integration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [operacional, atencao, erro, inativo]
        config:
          type: object
        lastSyncAt:
          type: string
          format: date-time
        errorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...